+++ internal/user/api.go
@@ -1,0 +1,156 @@
+package user
+
+import (
+	"context"
+	"fmt"
+	"github.com/openimsdk/openim-sdk-core/v3/internal/util"
+	"github.com/openimsdk/openim-sdk-core/v3/pkg/constant"
+	"github.com/openimsdk/openim-sdk-core/v3/pkg/db/model_struct"
+	"github.com/openimsdk/openim-sdk-core/v3/pkg/sdkerrs"
+	"github.com/openimsdk/openim-sdk-core/v3/pkg/utils"
+	"github.com/openimsdk/protocol/sdkws"
+	userPb "github.com/openimsdk/protocol/user"
+	"github.com/openimsdk/tools/log"
+)
+
+// GetSingleUserFromSvr retrieves user information from the server.
+func (u *User) GetSingleUserFromSvr(ctx context.Context, userID string) (*model_struct.LocalUser, error) {
+	users, err := u.GetUsersInfoFromSvr(ctx, []string{userID})
+	if err != nil {
+		return nil, err
+	}
+	if len(users) > 0 {
+		return users[0], nil
+	}
+	return nil, sdkerrs.ErrUserIDNotFound.WrapMsg(fmt.Sprintf("getSelfUserInfo failed, userID: %s not exist", userID))
+}
+
+// getSelfUserInfo retrieves the user's information.
+func (u *User) getSelfUserInfo(ctx context.Context) (*model_struct.LocalUser, error) {
+	userInfo, errLocal := u.GetLoginUser(ctx, u.loginUserID)
+	if errLocal != nil {
+		srvUserInfo, errServer := u.GetServerUserInfo(ctx, []string{u.loginUserID})
+		if errServer != nil {
+			return nil, errServer
+		}
+		if len(srvUserInfo) == 0 {
+			return nil, sdkerrs.ErrUserIDNotFound
+		}
+		userInfo = ServerUserToLocalUser(srvUserInfo[0])
+		_ = u.InsertLoginUser(ctx, userInfo)
+	}
+	return userInfo, nil
+}
+
+// ProcessUserCommandGetAll get user's choice
+func (u *User) ProcessUserCommandGetAll(ctx context.Context) ([]*userPb.CommandInfoResp, error) {
+	localCommands, err := u.DataBase.ProcessUserCommandGetAll(ctx)
+	if err != nil {
+		return nil, err // Handle the error appropriately
+	}
+
+	var result []*userPb.CommandInfoResp
+	for _, localCommand := range localCommands {
+		result = append(result, &userPb.CommandInfoResp{
+			Type:       localCommand.Type,
+			CreateTime: localCommand.CreateTime,
+			Uuid:       localCommand.Uuid,
+			Value:      localCommand.Value,
+		})
+	}
+
+	return result, nil
+}
+
+func (u *User) UserOnlineStatusChange(users map[string][]int32) {
+	for userID, onlinePlatformIDs := range users {
+		status := userPb.OnlineStatus{
+			UserID:      userID,
+			PlatformIDs: onlinePlatformIDs,
+		}
+		if len(status.PlatformIDs) == 0 {
+			status.Status = constant.Offline
+		} else {
+			status.Status = constant.Online
+		}
+		u.listener().OnUserStatusChanged(utils.StructToJsonString(&status))
+	}
+}
+
+func (u *User) GetUsersInfo(ctx context.Context, userIDs []string) ([]*model_struct.LocalUser, error) {
+	return u.GetUsersInfoFromSvr(ctx, userIDs)
+}
+
+func (u *User) GetSelfUserInfo(ctx context.Context) (*model_struct.LocalUser, error) {
+	return u.getSelfUserInfo(ctx)
+}
+
+// Deprecated: user SetSelfInfoEx instead
+func (u *User) SetSelfInfo(ctx context.Context, userInfo *sdkws.UserInfo) error {
+	userInfo.UserID = u.loginUserID
+	if err := u.updateSelfUserInfo(ctx, userInfo); err != nil {
+		return err
+	}
+	_ = u.SyncLoginUserInfo(ctx)
+	return nil
+}
+func (u *User) SetSelfInfoEx(ctx context.Context, userInfo *sdkws.UserInfoWithEx) error {
+	userInfo.UserID = u.loginUserID
+	if err := u.updateSelfUserInfoEx(ctx, userInfo); err != nil {
+		return err
+	}
+	_ = u.SyncLoginUserInfo(ctx)
+	return nil
+}
+func (u *User) SetGlobalRecvMessageOpt(ctx context.Context, opt int) error {
+	if err := util.ApiPost(ctx, constant.SetGlobalRecvMessageOptRouter,
+		&userPb.SetGlobalRecvMessageOptReq{UserID: u.loginUserID, GlobalRecvMsgOpt: int32(opt)}, nil); err != nil {
+		return err
+	}
+	err := u.SyncLoginUserInfo(ctx)
+	if err != nil {
+		log.ZWarn(ctx, "SyncLoginUserInfo", err)
+	}
+	return nil
+}
+
+func (u *User) UpdateMsgSenderInfo(ctx context.Context, nickname, faceURL string) (err error) {
+	if nickname != "" {
+		if err = u.DataBase.UpdateMsgSenderNickname(ctx, u.loginUserID, nickname, constant.SingleChatType); err != nil {
+			return err
+		}
+	}
+	if faceURL != "" {
+		if err = u.DataBase.UpdateMsgSenderFaceURL(ctx, u.loginUserID, faceURL, constant.SingleChatType); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// ProcessUserCommandAdd CRUD user command
+func (u *User) ProcessUserCommandAdd(ctx context.Context, userCommand *userPb.ProcessUserCommandAddReq) error {
+	err := u.processUserCommandAdd(ctx, userCommand)
+	if err != nil {
+		return err
+	}
+	return u.SyncAllCommand(ctx)
+}
+
+// ProcessUserCommandDelete delete user's choice
+func (u *User) ProcessUserCommandDelete(ctx context.Context, userCommand *userPb.ProcessUserCommandDeleteReq) error {
+	err := u.processUserCommandDelete(ctx, userCommand)
+	if err != nil {
+		return err
+	}
+	return u.SyncAllCommand(ctx)
+}
+
+// ProcessUserCommandUpdate update user's choice
+func (u *User) ProcessUserCommandUpdate(ctx context.Context, userCommand *userPb.ProcessUserCommandUpdateReq) error {
+	err := u.processUserCommandUpdate(ctx, userCommand)
+	if err != nil {
+		return err
+	}
+	return u.SyncAllCommand(ctx)
+}
